Major rebuild using Docker over Vagrant
This changes the build scripts to be simpler (use yum-builddep to pull
in most dependencies) and far, far faster because it's using Docker and
containers and UnionFS and all the joy that brings.

In addition, this adds support for CI builds on Travis and cleans up old
files.  Double yay!

Finally, this fixes #2 because all dependencies are more-or-less
automatically installed. Triple yay!